Refactored the namespace. AssetBundleCreator is now ModelCreator
