+++ Python/tdw/tdw_utils.py
@@ -1037,3 +1037,35 @@ class TDWUtils:
                 raise Exception(s)
         return {"x": monitor.x + monitor.width // 2 - window_width // 2,
                 "y": monitor.y + monitor.height // 2 - window_height // 2 + title_bar_height}
+
+
+    @staticmethod
+    def get_path(path: Union[str, Path]) -> Path:
+        """
+        :param path: A path as either a string or a `Path`.
+
+        :return: The path as a `Path`.
+        """
+
+        if isinstance(path, str):
+            return Path(path)
+        elif isinstance(path, Path):
+            return path
+        else:
+            raise Exception(path)
+
+    @staticmethod
+    def get_string_path(path: Union[str, Path]) -> str:
+        """
+        :param path: A path as either a string or a `Path`.
+
+        :return: The path as a string.
+        """
+
+        if isinstance(path, str):
+            p = path
+        elif isinstance(path, Path):
+            p = str(path.resolve())
+        else:
+            raise Exception(path)
+        return p.replace("\\", "/")
